GitHub/project_lion/liongram/requirements/local.txt
-r base.txt
django-debug-toolbar==3.3.0
